working on separate card component
